requirements.txt
python-telegram-bot>=21.0,<22.0
Pillow>=10.0.0
pyphen>=0.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
